chore: Update npm dependencies and add new packages
This commit updates the npm dependencies in the package.json file. It adds the following packages:
- @ai-sdk/openai@0.0.36
- ai@3.2.19
- zod@3.23.8

Additionally, it updates the devDependencies by adding eslint@8 and eslint-config-next@14.2.4.

These changes were made to ensure that the project is using the latest stable versions of the dependencies and to incorporate new functionality provided by the added packages.